Be more flexible in separating ingredient name and preparation
There are two goals here. First, it turns out that it can be hard to
type `--` on some systems. For example, on iOS, typing `--` will get
autocorrected to an em dash. By accepting an em dash as a separator, we
make it much more feasible to actually edit recipes on mobile devices.

Second, we can potentially parse out more different ways the preparation
note might be separated when importing recipes from external sources.