Handle multiple amounts and units (#7)
Add in better, but not perfect, handling for multiple amounts in an ingredient
line. 

### Notable Failures

This doesn't handle ingredients and amounts that have a space in them, which
the BBC website uses quite a bit:

```
600ml/20fl oz
```

This PR will correct remove read the `600ml` and remove the `20fl` from the
parsing, but it will not remove the `oz`. Some playing around with trying to
also remove the oz resulted in regressions in other tests, so I decided to not
pursue this further for now. I think a better solution will be a combination of
amount/unit parsing that is greedier, but also smarter.

I'm tracking this new issue in #6.

Resolve #4